# feat: static pages