--- a/front/pages/home/about.tsx
+++ b/front/pages/home/about.tsx
@@ -27,7 +27,7 @@ import {
 } from "@app/components/home/Particles";
 import { classNames } from "@app/lib/utils";
 
-export async function getServerSideProps() {
+export async function getStaticProps() {
   return {
     props: {
       shape: getParticleShapeIndexByName(shapeNames.icosahedron),