--- a/front/pages/home/security.tsx
+++ b/front/pages/home/security.tsx
@@ -17,7 +17,7 @@ import {
 } from "@app/components/home/Particles";
 import { classNames } from "@app/lib/utils";
 
-export async function getServerSideProps() {
+export async function getStaticProps() {
   return {
     props: {
       shape: getParticleShapeIndexByName(shapeNames.icosahedron),