--- a/front/pages/home/solutions/dust-platform.tsx
+++ b/front/pages/home/solutions/dust-platform.tsx
@@ -16,7 +16,7 @@ import {
 } from "@app/components/home/Particles";
 import { classNames } from "@app/lib/utils";
 
-export async function getServerSideProps() {
+export async function getStaticProps() {
   return {
     props: {
       shape: getParticleShapeIndexByName(shapeNames.galaxy),