## Description
Use `getStaticProps` instead of `getServerSideProps` for all `/home/*` route.
The server side rendering is not needed as the methods just return static props. Even the gtm tracking id is setup at build time.
This improve loading speed, and should reduce cpu usage for those route (but that's not probably a significant improvement as I'm not sure they're that popular)

## Tests
- Not tested locally as there is no difference between getStaticProps and getServerSideProps in dev.
- Tested on front edge
- Ran lighthouse before and after, i.e [https://dust.tt/home/security](https://dust.tt/home/security)
  -  before: 
     - perf: 96
     - FCP: 0.5s
     - LCP: 1.3s
  - after:
    - perf: 99
    - FCP: 0.6s
    - LCP: 0.9s

## Risk
Low, easy to revert and only on the public website part

## Deploy Plan
- Deploy front